Backend/requirements.txt
Flask==2.2.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
requests==2.28.1
grpcio==1.51.1
SimpleJSON
flask-cors
Flask-PyMongo
selenium